Unref the worker object before quiting the mainloop
This commit protects against a subsequent SIGTERM which might arrive
after quiting the mainloop but before finalizing the worker object. The
effect is gdm-session-worker exiting without waiting for gdm-x-session,
which leads to the main process exiting before gdm-x-session finishes.

Since systemd relies on the SIGCHLD from the main gdm process to tell
when the service has stopped, this behavior breaks any unit that has a
Conflicts=gdm.service entry.